requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
pdfplumber==0.11.4
python-pptx==1.0.2
google-generativeai==0.8.3
huggingface_hub==0.24.6
googletrans==4.0.0rc1
pydantic==2.9.2
Pillow>=10.3.0
streamlit
pdfplumber
python-docx
python-pptx
google-generativeai
googletrans==4.0.0-rc1
huggingface_hub
reportlab>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0